+++ R/cr_fundref.r
@@ -1,12 +1,13 @@
 #' Search the CrossRef Fundref API
 #'
 #' @export
-#'
+#' @family crossref
 #' @param dois Search by a single DOI or many DOIs.
 #' @template args
 #' @template moreargs
 #' @template cursor_args
 #' @template field_queries
+#' @template sorting
 #' @param facet (logical) Include facet results. Boolean or string with
 #' field to facet on. Valid fields are *, affiliation, funder-name,
 #' funder-doi, orcid, container-title, assertion, archive, update-type,
@@ -77,14 +78,19 @@
 #' ## query.container-title
 #' cr_funders('10.13039/100000001', works = TRUE,
 #'   flq = c(`query.container-title` = 'Ecology'))
+#' 
+#' # select only certain fields to return
+#' res <- cr_funders('10.13039/100000001', works = TRUE, 
+#'   select = c('DOI', 'title'))
+#' names(res$data)
 #' }
 `cr_funders` <- function(dois = NULL, query = NULL, filter = NULL,
   offset = NULL, limit = NULL,  sample = NULL, sort = NULL, order = NULL,
   facet=FALSE, works = FALSE, cursor = NULL, cursor_max = 5000,
-  .progress="none", flq = NULL, ...) {
+  .progress="none", flq = NULL, select = NULL, ...) {
 
   args <- prep_args(query, filter, offset, limit, sample, sort,
-                    order, facet, cursor, flq)
+                    order, facet, cursor, flq, select)
   if (length(dois) > 1) {
     res <- llply(dois, fundref_GET, args = args, works = works,
                  cursor = cursor, cursor_max = cursor_max, ...,
@@ -167,10 +173,10 @@
 `cr_funders_` <- function(dois = NULL, query = NULL, filter = NULL,
   offset = NULL, limit = NULL,  sample = NULL, sort = NULL, order = NULL,
   facet=FALSE, works = FALSE, cursor = NULL, cursor_max = 5000,
-  .progress="none", parse=FALSE, flq = NULL, ...) {
+  .progress="none", parse=FALSE, flq = NULL, select = NULL, ...) {
 
   args <- prep_args(query, filter, offset, limit, sample, sort, order,
-                    facet, cursor, flq)
+                    facet, cursor, flq, select)
   if (length(dois) > 1) {
     llply(dois, fundref_GET_, args = args, works = works,
           cursor = cursor, cursor_max = cursor_max, parse = parse, ...,