+++ R/cr_prefixes.r
@@ -1,13 +1,14 @@
 #' Search CrossRef prefixes
 #'
 #' @export
-#'
+#' @family crossref
 #' @param prefixes (character) Publisher prefixes, one or more in a vector or
 #' list. Required.
 #' @template args
 #' @template moreargs
 #' @template cursor_args
 #' @template field_queries
+#' @template sorting
 #' @param facet (logical) Include facet results. Boolean or string with
 #' field to facet on. Valid fields are *, affiliation, funder-name,
 #' funder-doi, orcid, container-title, assertion, archive, update-type,
@@ -75,15 +76,19 @@
 #' ## query.container-title
 #' cr_prefixes("10.1016", works = TRUE,
 #'   flq = c(`query.container-title` = 'Ecology'))
+#' 
+#' # select only certain fields to return
+#' res <- cr_prefixes("10.1016", works = TRUE, select = c('DOI', 'title'))
+#' names(res$data)
 #' }
 
 `cr_prefixes` <- function(prefixes, query = NULL, filter = NULL, offset = NULL,
   limit = NULL, sample = NULL, sort = NULL, order = NULL, facet=FALSE,
   works = FALSE, cursor = NULL, cursor_max = 5000, .progress="none",
-  flq = NULL, ...) {
+  flq = NULL, select = NULL, ...) {
 
   args <- prep_args(query, filter, offset, limit, sample, sort, order,
-                    facet, cursor, flq)
+                    facet, cursor, flq, select)
   if (length(prefixes) > 1) {
     res <- llply(prefixes, prefixes_GET, args = args, works = works,
                  cursor = cursor, cursor_max = cursor_max, ...,
@@ -137,12 +142,13 @@
 
 #' @export
 #' @rdname cr_prefixes
-`cr_prefixes_` <- function(prefixes, query = NULL, filter = NULL, offset = NULL,
-  limit = NULL, sample = NULL, sort = NULL, order = NULL, facet=FALSE, works = FALSE,
-  cursor = NULL, cursor_max = 5000, .progress="none", parse=FALSE, flq = NULL, ...) {
+`cr_prefixes_` <- function(prefixes, query = NULL, filter = NULL, 
+  offset = NULL, limit = NULL, sample = NULL, sort = NULL, order = NULL, 
+  facet=FALSE, works = FALSE, cursor = NULL, cursor_max = 5000, 
+  .progress="none", parse=FALSE, flq = NULL, select = NULL, ...) {
 
   args <- prep_args(query, filter, offset, limit, sample, sort, order,
-                    facet, cursor, flq)
+                    facet, cursor, flq, select)
   if (length(prefixes) > 1) {
     llply(prefixes, prefixes_GET_, args = args, works = works,
           cursor = cursor, cursor_max = cursor_max, parse = parse,