+++ R/cr_works.R
@@ -1,12 +1,13 @@
 #' Search CrossRef works (articles)
 #'
 #' @export
-#'
-#' @param dois Search by a single DOI or many DOIs.
+#' @family crossref
+#' @param dois Search by a single DOI or many DOIs.  Note that using this parameter at the same time as the `query`, `limit`, `select` or `flq` parameter will result in an error.
 #' @template args
 #' @template moreargs
 #' @template cursor_args
 #' @template field_queries
+#' @template sorting
 #' @param facet (logical) Include facet results. Boolean or string with
 #' field to facet on. Valid fields are *, affiliation, funder-name,
 #' funder-doi, orcid, container-title, assertion, archive, update-type,
@@ -111,17 +112,22 @@
 #' ## query.author and query.title
 #' res <- cr_works(query = "ecology",
 #'   flq = c(query.author = 'Smith', query.title = 'cell'))
+#' 
+#' # select only certain fields to return
+#' res <- cr_works(query = "NSF", select = c('DOI', 'title'))
+#' names(res$data)
 #' }
 
 `cr_works` <- function(dois = NULL, query = NULL, filter = NULL, offset = NULL,
   limit = NULL, sample = NULL, sort = NULL, order = NULL, facet=FALSE,
-  cursor = NULL, cursor_max = 5000, .progress="none", flq = NULL, ...) {
+  cursor = NULL, cursor_max = 5000, .progress="none", flq = NULL, 
+  select = NULL, ...) {
 
   if (cursor_max != as.integer(cursor_max)) {
     stop("cursor_max must be an integer", call. = FALSE)
   }
   args <- prep_args(query, filter, offset, limit, sample, sort, order,
-                    facet, cursor, flq)
+                    facet, cursor, flq, select)
 
   if (length(dois) > 1) {
     res <- llply(dois, cr_get_cursor, args = args, cursor = cursor,
@@ -131,10 +137,10 @@
     df <- tbl_df(bind_rows(res))
     #exclude rows with empty DOI value until CrossRef API supports
     #input validation
-    if (nrow(df[df$DOI == "", ]) > 0) {
+    if (nrow(df[df$doi == "", ]) > 0) {
       warning("only data with valid CrossRef DOIs returned",  call. = FALSE)
     }
-    df <- df[!df$DOI == "", ]
+    df <- df[!df$doi == "", ]
     list(meta = NULL, data = df, facets = NULL)
   } else {
     tmp <- cr_get_cursor(dois, args = args, cursor = cursor,
@@ -158,13 +164,14 @@
 #' @rdname cr_works
 `cr_works_` <- function(dois = NULL, query = NULL, filter = NULL, offset = NULL,
   limit = NULL, sample = NULL, sort = NULL, order = NULL, facet=FALSE,
-  cursor = NULL, cursor_max = 5000, .progress="none", parse=FALSE, flq = NULL, ...) {
+  cursor = NULL, cursor_max = 5000, .progress="none", parse=FALSE, 
+  flq = NULL, select = NULL, ...) {
 
   if (cursor_max != as.integer(cursor_max)) {
     stop("cursor_max must be an integer", call. = FALSE)
   }
   args <- prep_args(query, filter, offset, limit, sample, sort, order,
-                    facet, cursor, flq)
+                    facet, cursor, flq, select)
 
   if (length(dois) > 1) {
     llply(dois, cr_get_cursor_, args = args, cursor = cursor,
@@ -200,6 +207,7 @@ cr_get_cursor_ <- function(x, args, cursor, cursor_max, parse, ...) {
 }
 
 parse_meta <- function(x) {
+  if (is.null(x$message)) return(data.frame(NULL))
   tmp <- x$message[ !names(x$message) %in% c('facets','items') ]
   st <- tmp$query$`search-terms`
   data.frame(total_results = tmp$`total-results`,
@@ -283,13 +291,20 @@ parse_works <- function(zzz){
   } else if (all(is.na(zzz))) {
     NULL
   } else {
-    out_tmp <- data.frame(as.list(unlist(lapply(keys, manip, y = zzz))),
-                          stringsAsFactors = FALSE)
+    tmp <- unlist(lapply(keys, manip, y = zzz))
+    #tmp[vapply(tmp, function(z) nchar(z) == 0 || is.na(z), TRUE)] <- NULL
+    out_tmp <- data.frame(
+      as.list(Filter(function(x) nchar(x) > 0, tmp)), 
+      stringsAsFactors = FALSE)
+    # out_tmp <- data.frame(as.list(unlist(lapply(keys, manip, y = zzz))),
+    #                       stringsAsFactors = FALSE)
     out_tmp$assertion <- list(parse_todf(zzz$assertion)) %||% NULL
     out_tmp$author <- list(parse_todf(zzz$author)) %||% NULL
     out_tmp$funder <- list(parse_todf(zzz$funder)) %||% NULL
     out_tmp$link <- list(parse_todf(zzz$link)) %||% NULL
     out_tmp$`clinical-trial-number` <- list(parse_todf(zzz$`clinical-trial-number`)) %||% NULL
+    out_tmp <- Filter(function(x) length(unlist(x)) > 0, out_tmp)
+    names(out_tmp) <- tolower(names(out_tmp))
     return(out_tmp)
   }
 }