+++ man-roxygen/args.R
@@ -1,5 +1,8 @@
 #' @param query Query terms
-#' @param filter Filter options. See details.
+#' @param filter Filter options. See examples for usage examples
+#' and \code{\link{filters}} for what filters are available.
+#' \code{filter} is available for use with \code{cr_works} and 
+#' other \code{crossref} family functions with \code{works=TRUE}
 #' @param offset Number of record to start at. Minimum: 1. For
 #' \code{\link{cr_works}}, and any function setting \code{works = TRUE},
 #' the maximum offset value is 10000. For larger requests use \code{cursor}.
@@ -8,6 +11,6 @@
 #' @param sample (integer) Number of random results to return. when you use
 #' the sample parameter, the rows and offset parameters are ignored.
 #' Ignored unless \code{works=TRUE}. Max: 100
-#' @param sort (character) Field to sort on, one of score, relevance, updated,
-#' deposited, indexed, or published.
 #' @param order (character) Sort order, one of 'asc' or 'desc'
+#' @param select (character) One or more field to return (only those fields 
+#' are returned)