+++ man-roxygen/field_queries.R
@@ -1,12 +1,15 @@
 #' @param flq field queries. One or more field queries. Acceptable set of 
 #' field query parameters are: 
 #' \itemize{
-#'  \item query.title	- Query title and subtitle
-#'  \item query.container-title	- Query container-title aka. publication name
-#'  \item query.author - Query author first and given names
-#'  \item query.editor - Query editor first and given names
-#'  \item query.chair	- Query chair first and given names
-#'  \item query.translator - Query translator first and given names
-#'  \item query.contributor	- Query author, editor, chair and translator 
+#'  \item \code{query.title} - Query title and subtitle
+#'  \item \code{query.container-title}	- Query container-title aka. publication name
+#'  \item \code{query.author} - Query author first and given names
+#'  \item \code{query.editor} - Query editor first and given names
+#'  \item \code{query.chair}	- Query chair first and given names
+#'  \item \code{query.translator} - Query translator first and given names
+#'  \item \code{query.contributor} - Query author, editor, chair and translator 
 #'  first and given names
+#'  \item \code{query.bibliographic} - Query bibliographic information, useful for 
+#'  citation lookup. Includes titles, authors, ISSNs and publication years
+#'  \item \code{query.affiliation} - Query contributor affiliations
 #' }