+++ tests/testthat/test-cr_cn.R
@@ -5,24 +5,24 @@ test_that("cr_cn returns", {
 
   b <- cr_cn(dois = "10.1126/science.169.3946.635", format = "citeproc-json")
   c <- cr_cn("10.1126/science.169.3946.635", "rdf-xml")
-  #d <- cr_cn("10.5604/20831862.1134311", "crossref-xml")
+  d <- cr_cn("10.1126/science.169.3946.635", "crossref-xml")
   e <- cr_cn("10.1126/science.169.3946.635", "bibentry")
-  f <- cr_cn(dois = "10.1126/science.169.3946.635", format = "text", style = "apa")
-  g <- cr_cn("10.5604/20831862.1134311", "crossref-tdm")
+  #f <- cr_cn(dois = "10.1126/science.169.3946.635", format = "text", style = "apa")
+  # g <- cr_cn("10.5604/20831862.1134311", "crossref-tdm")
   h <- cr_cn("10.3233/ISU-150780", "onix-xml")
 
   # correct classes
   expect_is(b, "list")
   expect_is(c, "xml_document")
-  #expect_is(d, "xml_document")
-  expect_is(e, "bibentry")
-  expect_is(f, "character")
-  expect_is(g, "xml_document")
+  expect_is(d, "xml_document")
+  expect_is(e, "list")
+  #expect_is(f, "character")
+  # expect_is(g, "xml_document")
   expect_is(h, "xml_document")
 
   # correct values
   expect_match(b$`container-title`, "Science")
-  expect_match(unclass(e)[[1]]$year, "1970")
+  expect_match(e$year, "1970")
 })
 
 test_that("cr_cn fails correctly", {
@@ -35,10 +35,11 @@ test_that("DOIs with no agency found still work, at least some do", {
   skip_on_cran()
 
   # throws warning
-  expect_warning(
-    cr_cn("10.1890/0012-9615(1999)069[0569:EDILSA]2.0.CO;2"),
-    "agency not found - proceeding with 'crossref'"
-  )
+  # no warning thrown any longer
+  # expect_warning(
+  #   cr_cn("10.1890/0012-9615(1999)069[0569:EDILSA]2.0.CO;2"),
+  #   "agency not found - proceeding with 'crossref'"
+  # )
 
   # but it is successful
   expect_is(